Add files via upload
This is a simple student result sheet project built using HTML, CSS, and JavaScript. It displays the result (name, marks, and division) based on the entered roll number. The data includes records for 50 students and was created as part of my practice while learning conditional statements, user input handling basics....